images/airflow/2.9.2/requirements/requirements.txt
--constraint /dev/null

dbt-core==1.9.4
dbt-extractor==0.6.0
dbt-snowflake==1.9.2
airflow-dbt-python==3.1.0
apache-airflow-providers-snowflake>=5.0.0,<5.2.0
apache-airflow-providers-sftp>=4.6.0,<4.8.0
snowflake-connector-python<4.0.0,>=3.13.1
apache-airflow-providers-slack>=4.0.0,<8.5.0
ndg-httpsclient==0.5.1
pyasn1==0.5.0
apache-airflow-providers-amazon>=8.0.0
boto3==1.34.128
httpx>=0.24.1,<1.0.0
tenacity>=8.2.0,<9.0.0



